requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
requests==2.31.0
beautifulsoup4==4.12.2
langcodes==3.3.0
language_data==1.3.0
dicttoxml==1.7.16
natsort==8.4.0
pyzbar==0.1.9
Pillow==11.0.0
numpy==1.26.4
cloudscraper==1.2.71
